Extend the 'Put PC in standby when TV is switched off' behaviour.
bugzid: 660

This replaces 'Put this PC in standby mode when the TV is switched off'
with an enum that allows the user to chose between 'Ignore', 'Suspend'
and 'Shutdown'.

Ideally this would co-ordinate with the global config and only offer,
for example, suspend, when it is actually available - but we are
building the settings screen programmatically from an xml file, so it's
not obvious how that might be done.